// vaccineApp/src/infrastructure/interfaces/dependent-bd-interface.tsx
// Generated by https://quicktype.io

export interface DependentBDResponse {
    dependents: DependentBD[];
    desde:      number;
    limite:     number;
    total:      number;
}

export interface DependentBD {
    _id:        ID;
    age:        number;
    birth:      string;
    days_birth: number;
    email:      string;
    gender_id:  string;
    isChildren: boolean;
    isUser:     boolean;
    lastname:   string;
    name:       string;
    phone:      string;
    status:     boolean;
    token:      string;
    user_id:    ID;
    state?:      string;
     city?:       string;
}

export interface ID {
    $oid: string;
}


// Generated by https://quicktype.io

export interface DependentByIDResponse {
    result: Result;
}

export interface Result {
    _id:        ID;
    name:       string;
    lastname:   string;
    email:      string;
    state:      string;
    phone?:      string;
    token?:      string;
    birth:      string;
    gender_id:  string;
    status:     boolean;
    isUser:     boolean;
    user_id:    ID;
    isChildren: boolean;
    age:        number;
    days_birth: number;
}

export interface ID {
    $oid: string;
}